LayoutTests/fast/js/sputnik/Conformance/12_Statement/12.9_The_return_Statement/S12.9_A1_T4-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Parse error
S12.9_A1_T4

PASS Expected parsing failure

TEST COMPLETE
